spi: add loopback test driver to allow for spi_master regression tests
This driver is submitting lots of distinct spi-messages messages
with all kinds of alignments and length pattern.
Also distinct kinds of transfer pattern tests are implemented
(rx, tx, rx/tx, tx+tx, tx+rx,...)

Right now on a raspberry pi 752 distinct spi_messages are executed
in 13 different scenarios.

Configuration of additional test-pattern is easy, so that when
new bugs in drivers get detected the relevant transfer pattern can
also get added to the test framework, so that such situations are
detected in other drivers as well.

The idea behind this driver is to make it possible to also detect
regressions in spi_master implementations when changes occur.
Potentially these tests could get executed automatically in a
test-server-farm.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>

Note that we do some rescheduling and resubmit the spi_message
when we see timeouts on spi_sync.

This is because under some circumstances (at least on the rpi) too
much IO is happening (with IOWAIT) that is triggering these
spi-timeouts (e.g: spi is scheduling a dma transfer and then waits
for completion, but the dma_engine worker thread does not get
woken because of the system load).

The chances of IO are higher because the test driver is reporting
on the tests it is executing to dmesg, which results in more disk-io.