dmaengine: bcm2835: move controlblock chain generation into separate method
In preparation of adding slave_sg functionality this patch moves the
generation/allocation of bcm2835_desc and the building of
the corresponding DMA-control-block chain from bcm2835_dma_prep_dma_cyclic
into the newly created method bcm2835_dma_create_cb_chain.

This new code also takes the limits of LITE channels into account
and splits the control-block transfers at the correct location.

Tested Audio output with a Hifiberry I2S card.

Note that the bcm2835-i2s driver that is currenlty in the kernel
fails to work since commit 94cb7f76caa0b337 and requires
modifications to use the new clock-driver framework.
For this testing a quick and dirty patch had to be used - a
patch to fix this is in the pipeline.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>