ASoC: bcm2835: move to use the clock framework
Since the move to the new clock framework with
commit 94cb7f76caa0b337 ("Switch to using the new clock driver support")
this driver was no longer functional as it was manipulating the
clock registers locally without going true the framework.

This patch moves to use the new clock framework and also
moves away from the hardcoded address offsets for DMA getting
the dma-address directly from the device tree.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>