memory: bcm2835: add register information and debugfs reginfo support
Add general register information and the corresponding
reginfo file in debugfs to dump the registers.

Unfortunately the firmware seems to restrict access to the
sd-HW block by mapping the access to a zero page.
this results in only being able to read 0.

So as of now this is just for completeness - if the
firmware changes its mind we can read those registers.

Note that the whole 0x7eeXXXXX range seems impacted by this
maybe also 0x7efXXXXX.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>