can: mcp25xxfd: compute CRC for alternative bit patterns
The mcp2517 has an issue with regards to CRC computation in
conjunction with a "late" bit flip of bit 7 of any register byte due
to activity on the CAN bus (transmission or reception just finished
before the register byte is transferred)

This can result in the wrong crc transmitted by the controller.

So we try to compute the CRC of the possibile variants and if one of
them matches use that one.

As a fallback we still can reread the relevant registers.

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>